backend/requirements.txt
# Основные
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.5
Pillow==10.0.1
djoser==2.2.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
reportlab==4.0.4
python-dotenv==1.0.0
gunicorn==21.2.0
drf_extra_fields==3.4.0
drf-spectacular==0.27.1

# Для разработки (dev-requirements.txt)
flake8==6.1.0
django-import-export==3.3.4